requirements.txt
python-dotenv==1.0.1
cryptography>=44.0.1
PySide6==6.7.1
pyotp
qrcode
Flask
pywebview
qtpy
flask-talisman
flask-seasurf
Flask-Limiter
pynacl
boto3